refactor all deletes methods and file names to remove
