+++ backend/server/controllers/index.js
@@ -4,5 +4,5 @@ exports.get = require('./get');
 exports.post = require('./post');
 exports.put = require('./put');
 exports.patch = require('./patch');
-exports.deletes = require('./deletes');
+exports.remove = require('./remove');
