+++ backend/server/models/index.js
@@ -4,4 +4,4 @@ exports.get = require('./get.js');
 exports.post = require('./post.js');
 exports.put = require('./put.js');
 exports.patch = require('./patch.js');
-exports.deletes = require('./deletes.js');
+exports.remove = require('./remove.js');
